Explicit cast

+++ src/libsodium/crypto_core/ed25519/ref10/ed25519_ref10.c
@@ -529,7 +529,7 @@ equal(signed char b, signed char c)
     unsigned char ub = b;
     unsigned char uc = c;
     unsigned char x  = ub ^ uc; /* 0: yes; 1..255: no */
-    uint32_t      y  = x;       /* 0: yes; 1..255: no */
+    uint32_t      y  = (uint32_t) x; /* 0: yes; 1..255: no */
 
     y -= 1;   /* 4294967295: yes; 0..254: no */
     y >>= 31; /* 1: yes; 0: no */
